+++ fractal/components/assessments/RecommendationSummaryGroup/RecommendationSummaryGroup.config.js
@@ -1,15 +1,13 @@
 const summaryConfig = require('../RecommendationSummary/RecommendationSummary.config');
 const summaryScoresConfig = require('../RecommendationScoreSummary/RecommendationScoreSummary.config');
 
-const getModuleConfigs = () => {
-    return [
+const config = [
         {title: 'Your MoneyFit Score', uri: 'moneyfitScore', internalName: 'moneyfitScore', summaryType: 'overall'},
         {title: 'Liquidity', uri: 'liquidity', internalName: 'liquidity'},
         {title: 'Debt-to-Income Ratio', uri: 'debt_ratio', internalName: 'debtRatio'},
         {title: 'Savings', uri: 'savings', internalName: 'savings'},
         {title: 'Life Insurance', uri: 'life_insurance', internalName: 'lifeInsurance'}
     ];
-};
 
 module.exports = {
     "title": "Assessment Recommendation Summary Group",
@@ -17,34 +15,13 @@ module.exports = {
     "collated": true,
     "preview": "@recommendation-summary-group-preview",
     "context": {
-        getModuleConfigs: getModuleConfigs,
-        getData: function(recommendationModuleUri) {
-            let module = summaryConfig.variants[0].context;
-            switch (recommendationModuleUri) {
-                case 'moneyfitScore':
-                    module = summaryScoresConfig.context;
-                    break;
-                case 'liquidity':
-                    module = summaryConfig.variants[1].context;
-                    break;
-                case 'debt_ratio':
-                    module = summaryConfig.variants[2].context;
-                    break;
-                case 'savings':
-                    module = summaryConfig.variants[3].context;
-                    break;
-                case 'life_insurance':
-                    module = summaryConfig.variants[4].context;
-                    break;
-                default:
-                    break;
-            }
-            return new Promise((resolve, reject) => {
-                resolve(module);
-            });
-        },
-        getDataSuccess: function(recommendation) {
-            return recommendation;
+        baseUrl: 'somewhere',
+        recommendations: {
+            moneyfitScore: Object.assign({}, summaryScoresConfig.context, {config: config[0]}),
+            liquidity: Object.assign({}, summaryConfig.variants[1].context, {config: config[1]}),
+            debt_ratio: Object.assign({}, summaryConfig.variants[2].context, {config: config[2]}),
+            savings: Object.assign({}, summaryConfig.variants[3].context, {config: config[3]}),
+            life_insurance: Object.assign({}, summaryConfig.variants[4].context, {config: config[4]})
         }
     }
 };