+++ fractal/components/assessments/RecommendationSummaryView/RecommendationSummaryView.config.js
@@ -1,0 +1,9 @@
+const summaryConfig = require('../RecommendationSummaryGroup/RecommendationSummaryGroup.config');
+
+module.exports = {
+    "title": "Assessment Recommendation Summary View",
+    "status": "ready",
+    "collated": true,
+    "preview": "@recommendation-summary-view-preview",
+    "context": summaryConfig.context
+};